feat: move refactored test back
