+++ util/logutil/hex_test.go
@@ -17,48 +17,52 @@ import (
 	"bytes"
 	"encoding/hex"
 	"reflect"
+	"testing"
 
-	. "github.com/pingcap/check"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/util/logutil"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/suite"
 )
 
-var _ = Suite(&testHexSuite{})
+type HexTestSuite struct {
+	suite.Suite
+}
 
-type testHexSuite struct{}
-
-func (s *testHexSuite) SetUpSuite(c *C) {}
-
-func (s *testHexSuite) SetUpTest(c *C) {}
-
-func (s *testHexSuite) TestHex(c *C) {
+func (s *HexTestSuite) TestHex() {
+	assert := assert.New(s.T())
 	var region metapb.Region
 	region.Id = 6662
 	region.StartKey = []byte{'t', 200, '\\', 000, 000, 000, '\\', 000, 000, 000, 37, '-', 000, 000, 000, 000, 000, 000, 000, 37}
 	region.EndKey = []byte("3asg3asd")
 
-	c.Assert(logutil.Hex(&region).String(), Equals, "{Id:6662 StartKey:74c85c0000005c000000252d0000000000000025 EndKey:3361736733617364 RegionEpoch:<nil> Peers:[] EncryptionMeta:<nil>}")
+	assert.Equal(logutil.Hex(&region).String(), "{Id:6662 StartKey:74c85c0000005c000000252d0000000000000025 EndKey:3361736733617364 RegionEpoch:<nil> Peers:[] EncryptionMeta:<nil>}")
 }
 
-func (s *testHexSuite) TestPrettyPrint(c *C) {
+func (s *HexTestSuite) TestPrettyPrint() {
+	assert := assert.New(s.T())
 	var buf bytes.Buffer
 
 	byteSlice := []byte("asd2fsdafs中文3af")
 	logutil.PrettyPrint(&buf, reflect.ValueOf(byteSlice))
-	c.Assert(buf.String(), Equals, "61736432667364616673e4b8ade69687336166")
-	c.Assert(buf.String(), Equals, hex.EncodeToString(byteSlice))
+	assert.Equal(buf.String(), "61736432667364616673e4b8ade69687336166")
+	assert.Equal(buf.String(), hex.EncodeToString(byteSlice))
 	buf.Reset()
 
 	// Go reflect can't distinguish uint8 from byte!
 	intSlice := []uint8{1, 2, 3, uint8('a'), uint8('b'), uint8('c'), uint8('\'')}
 	logutil.PrettyPrint(&buf, reflect.ValueOf(intSlice))
-	c.Assert(buf.String(), Equals, "01020361626327")
+	assert.Equal(buf.String(), "01020361626327")
 	buf.Reset()
 
 	var ran kv.KeyRange
 	ran.StartKey = kv.Key("_txxey23_i263")
 	ran.EndKey = nil
 	logutil.PrettyPrint(&buf, reflect.ValueOf(ran))
-	c.Assert(buf.String(), Equals, "{StartKey:5f747878657932335f69323633 EndKey:}")
+	assert.Equal(buf.String(), "{StartKey:5f747878657932335f69323633 EndKey:}")
+}
+
+func TestHexTestSuite(t *testing.T) {
+	suite.Run(t, new(HexTestSuite))
 }
